profile.c: Avoid warning:
profile.c:1212: warning: comparison between signed and unsigned

git-svn-id: https://www.opensc-project.org/svnp/opensc/trunk@5373 c6295689-39f2-0310-b995-f0e70906c6a9

+++ src/pkcs15init/profile.c
@@ -1209,7 +1209,7 @@ new_file(struct state *cur, const char *name, unsigned int type)
 		profile->df[df_type] = file;
 	}
 	assert(file);
-	if (file->type != (int)type) {
+	if (file->type != type) {
 		parse_error(cur, "inconsistent file type (should be %s)",
 			file->type == SC_FILE_TYPE_DF 
 				? "DF" : file->type == SC_FILE_TYPE_BSO
